armv8: fsl-lsch2: disable ECC on armv8 chassis 2 platforms
The Read DMA operations get early termination indication from the
controller. This issue is observed as CRC error in the status
registers. The issue is due to address collision at address 0 in
the dual port memory. The read is a dummy read to flush out the
header, but due to collision the controller logs the mbit error
reported by the ECC check logic. This results in the early termination
of the Read DMA operation by the controller. The issue happens to
all the interface speeds(GEN1/2/3) for all the products.

In order to enable SATA, ECC needs to be disabled by writing  the data
0x80000000 to the address 0x20140520.

Signed-off-by: Tang Yuantian <yuantian.tang@nxp.com>