lx2160aqds: describe MDIO bus routing and phy devices
This patch describes the MDIO bus routing to various IO slots and
various phy devices attached to the MDIO bus.

A table is maintained to store above information for various Reset
Control Word (RCW) protocols. This table can be overridden using
"dpmac_override" in u-boot environment variable hwconfig.

In this scenerio use "dpmac" environment variable to detect the PHYs
attached to a dpmac and MDIO bus to use for reading PHY registers and
IO slot to route the MDIO bus to.

By default the environment variable "dpmac" has been defined for
serdes configuration 19_5_2.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>