Arm64:lx2160a:PCIE: Modify the inbound window set.
Due to the MSIX talbe store the BAR1 for all VFs and PFs, so no
need to set the inbound window for BAR1 for all VFs and PFs, only
need enable the inbound window.

Due to the u-boot don't support the SRIOV feature, so don't need
set the inbound window for all VFs.

Signed-off-by: Bao Xiaowei <xiaowei.bao@nxp.com>