fsl/lx2160aqds: fixup the dpmac node to specify phy mode
The 2 RGMIIs on the LX2 QDS board need to be configured as
PHY_INTERFACE_MODE_RGMII_ID in order to enable the TX internal
delay. At the moment, MC does not distinguish between RGMII, RGMII_ID
or the other variants. The 'phy-connection-type' attribute is used
to specify the exact interface mode (rgmii-id) so that the dpaa2-mac
driver can correctly connect to the phydev.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>