pci: lx_pcie: Add WA for LX2 PCIe is'nt sending ACK in required time
When LX2 PCIe controller is sending multiple split completions and
ACK latency expires indicating that ACK should be send at priority.
But because of large number of split completions and FC update DLLP,
the controller does not give priority to ACK transmission. This
results into ACK latency timer timeout error at the link partner and
the pending TLPs are replayed by the link partner again.

Workaround:
1. Reduce the ACK latency timeout value to a very small value.
2. Restrict the number of completions from the LX2 PCIe controller
   to 1, by changing the Max Read Request Size (MRRS) of link partner
   to the same value as Max Packet size (MPS).

Note:
The part 1 of the Workaround has been implemented by commit:
3ddd829ee PCI: lx_pcie: fix the ACK latency timeout
This patch is for part 2.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>