LA1224: Add support for LA1224 board
Added support for AQR113C
Disable SGMII 17 & 18 even if serdes protocol is enabling it
because these are not used.
Enable RGMII 17 in place of SGMII17 and disable RGMII18.
Deleted unnecessary code from board file for this board

This patch adds modifications used to test the board.
It is not a final patch.

Signed-off-by: Razvan Ionut Cirjan <razvanionut.cirjan@nxp.com>
Signed-off-by: Meenakshi Aggarwal <meenakshi.aggarwal@nxp.com>