board/freescale/vid: rework of VID support
The VID code does not properly set the voltages on all chips.
It does not properly reports voltage changes.
For IR, the read accuracy was incorrect and for
the LTC the step width was too wide.

This patchset fix above problems and also make LTC channel
use configurable to deal properly with single and dual
channel setups.

Signed-off-by: Dan Nica <dan.nica@nxp.com>
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Wasim Khan <wasim.khan@nxp.com>