LA1224: Complete the commit 6166ac6da37
Added support for RGMII 17 port.
Remove support for all ports that are not supported by this
by this board even though SerDes protocol used enable its.
Remove the configs and support for Inphi and Cortina retimers.
Remove the defines for AQR107.
Remove the support for SerDes1 protocol 19 and add 22.
Remove the register of Mdio2
QUBOOT-5673

Signed-off-by: Razvan Ionut Cirjan <razvanionut.cirjan@nxp.com>
Signed-off-by: Vabhav Sharma <vabhav.sharma@nxp.com>