lx2-uboot: Added support for BlueBox3 PCIe T6 Slot 6
On bootup, this code will detect inserted cards either by configuration
setting or by presence detect signals. It will reconfigure the
equalizer chips appropriately for the active lane width and issue a
reset to recover the PCIe link if the lane configuration has changed.

The reset is needed because no deterministic way is known to recover
the LTSSM after it has fallen into compliance testing mode. That can
happen if more lanes are active than the card actually has.

Great care is taken to ensure that the reset preserves the boot source
and the QSPI bank selection if possible to simplify usage.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>