arm64: layerscape: Move GIC RD tables initialization to setup function
Move GIC redistributor tables initialization to CPU setup function.

This patch introduces a GIC redistributor tables init function, and
moves the function of reserving memory for GIC redistributor tables
to soc.c and adds a argument for the memory size to reserve, BTW
rename the function so that it is more readable.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
[Backported]
Signed-off-by: Priyanka Jain <priyanka.jain@nxp.com>