lx2160a_common.h: modify "mcmemsize=0x70000000\0" to "mcmemsize=0x20000000\0"
When running the memory init function, the environment has not been loaded yet.
This results in the memory not being initialized with configured env variables.
Initializing with the default env variables, the memory passed by the U-boot to
the kernel will miss 2GB forever.
For fixing this issue, I modify the default value of mcmemsize to 0x20000000.

Signed-off-by: Jianchao Wang <jianchao.wang_1@nxp.com>