board: freescale: la1224ardb: choose DTS based on the SerDes protocol
The LA1224ARDB board supports on SerDes1 protocols 18 and 22. Depending
on the current SerDes protocol, we need to choose a different DTS which
properly describes the Ethernet ports.

Implement the board_fit_config_name_match() function so that the this is
dynamically chosen by u-boot without any user intervention or
modifications.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>