board: freescale: la1224ardb: fix retimer configuration
With DM_ETH enabled, the case for the LA1224ARDB board, the
board_eth_init() function is not called because all ethernet devices are
created based on the device-tree.

Because of this, the DS250DF230ZLST 25-Gbps TI retimer does not get
configured for the appropriate frequency depeding on the SerDes protocol
currently run on the LX2160A SoC.

Remove the board_eth_init() function entirely and instead add a new
function - board_retimer_init() - which will get called from the
board_init() function found in the lx2160a.c source file. The necessary
setup steps for the TI retimer will get done from this new function.

Also, the I2C address of the TI retimer is not 0x1, as previously
defined, but instead 0x18. Fix this as well.

At the same time, add the static keyword to some local functions.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>